WPF example updates; Update ID3v2 frame writer to detect Unicode and change TextEncoding byte

+++ Examples/IdSharp.Tagging.Harness.Wpf/ViewModel/Interfaces/IScanViewModel.cs
@@ -12,7 +12,7 @@ namespace IdSharp.Tagging.Harness.Wpf.ViewModel.Interfaces
         ICommand ScanCommand { get; }
         ICommand CancelCommand { get; }
         ICommand BrowseCommand { get; }
-        int PercentComplete { get; }
+        double PercentComplete { get; }
         ObservableCollection<Track> TrackCollection { get; }
     }
 }
